Wrap BitmapFactory.Decode methods with try/catch
There is case when android images are not released even when they have not references
In that case creation of new images throws OutOfMemory exception by Java runtime.
We catch that exception and force Mono to collect garbage. Which helped to release memory.
That fix verified on the emulator, because it has been reproduced with 100%.
Though it has not been proved on the device, becasue Mono collects garbage much often.
I assume the fix still should increase stability of the image loading